General code cleanup.

+++ library/sources/XLCellReference.cpp
@@ -252,7 +252,7 @@ uint16_t XLCellReference::columnAsNumber(const std::string& column)
     uint16_t result = 0;
 
     for (int16_t i = static_cast<int16_t>(column.size() - 1), j = 0; i >= 0; --i, ++j) {
-        result += static_cast<uint16_t>((column[i] - asciiOffset) * std::pow(alphabetSize, j));
+        result += static_cast<uint16_t>((column[static_cast<uint64_t>(i)] - asciiOffset) * std::pow(alphabetSize, j));
     }
 
     return result;
